+++ pmd-java/src/test/java/net/sourceforge/pmd/lang/java/ast/AllJavaAstTreeDumpTest.java
@@ -19,9 +19,10 @@ import org.junit.platform.suite.api.Suite;
     Java21TreeDumpTest.class,
     Java22TreeDumpTest.class,
     Java23TreeDumpTest.class,
-    Java23PreviewTreeDumpTest.class,
     Java24TreeDumpTest.class,
-    Java24PreviewTreeDumpTest.class
+    Java24PreviewTreeDumpTest.class,
+    Java25TreeDumpTest.class,
+    Java25PreviewTreeDumpTest.class
 })
 class AllJavaAstTreeDumpTest {